Added Dynamically Sized Table
Added dynamic table, though I think we should review whether we actually use a table to display our tasks. This is why I'm branching it for now. Currently it only works to remove from the bottom of the list up. But I think if we hammer down the way we are going to store and link data it shouldnt be too hard. I think it will make it harder to do things like drop down menus from a table than from separate children tasks. It should be considered heavily tbh.